backend/root/requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
gunicorn==20.1.0
marshmallow==3.19.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
pytest